Update `int` dtypes to precision capable ones

+++ pyFV3/stencils/fillz.py
@@ -5,7 +5,7 @@ from gt4py.cartesian.gtscript import BACKWARD, FORWARD, PARALLEL, computation, i
 import ndsl.dsl.gt4py_utils as utils
 from ndsl import QuantityFactory, StencilFactory, orchestrate
 from ndsl.constants import X_DIM, Y_DIM, Z_DIM
-from ndsl.dsl.typing import Float, FloatField, FloatFieldIJ, IntFieldIJ
+from ndsl.dsl.typing import Float, Int, FloatField, FloatFieldIJ, IntFieldIJ
 from pyFV3.tracers import Tracers
 
 
@@ -133,7 +133,7 @@ class FillNegativeTracerValues:
 
         # Setting initial value of upper_fix to zero is only needed for validation.
         # The values in the compute domain are set to zero in the stencil.
-        self._zfix = quantity_factory.zeros([X_DIM, Y_DIM], units="unknown", dtype=int)
+        self._zfix = quantity_factory.zeros([X_DIM, Y_DIM], units="unknown", dtype=Int)
         self._sum0 = quantity_factory.zeros(
             [X_DIM, Y_DIM],
             units="unknown",
